feat: initial curation complete
