+++ app/core/lib/curations/helpers/cli.ts
