+++ app/core/lib/curations/index.ts
@@ -1,0 +1,10 @@
+import type { CommandDefinition } from "~/core/lib/curations/types/curation";
+import { curationModules } from "./modules";
+
+interface CommandBusInterface {
+  nouns: CommandDefinition[];
+}
+
+export const commandBus: CommandBusInterface = {
+  nouns: curationModules.nouns.commandBus,
+};
