+++ app/core/lib/curations/stores/CurationAuditTrail.ts
@@ -1,4 +1,7 @@
-import type { CurationAudit, CurationObject } from "~/types/curation";
+import type {
+  CurationAudit,
+  CurationObject,
+} from "~/core/lib/curations/types/curation";
 
 import type { Firestore } from "firebase-admin/firestore";
