+++ app/core/lib/curations/types/config.ts
@@ -1,0 +1,27 @@
+import type {
+  CommandDefinition,
+  ContentGenerator,
+  CurationObject,
+  ReviewGenerator,
+} from "~/core/lib/curations/types/curation";
+
+import { CurationRepository } from "../stores/CurationRepository";
+
+export type CurationConfig<T extends CurationObject> = {
+  namespace: string;
+  repo: CurationRepository<T>;
+  generator: ContentGenerator<T>;
+  reviewer: ReviewGenerator<T>;
+  commandBus: CommandDefinition[];
+  label?: string;
+  description?: string;
+  generateRequestId?: (data: Partial<T>) => string;
+  cli?: {
+    printSummary?: (entry: T) => void;
+    printRow?: (entry: T) => void;
+  };
+  ui?: {
+    renderRow?: (entry: T) => React.ReactNode[];
+    renderDetail?: (entry: T) => React.ReactNode;
+  };
+};
